requirements.txt
Django==1.8.6
django-dual-authentication==1.0.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
httpie==0.9.3
Pygments==2.1.3
PyJWT==1.4.0
requests==2.10.0
